+++ readium/streamer/CHANGELOG.md
@@ -4,7 +4,9 @@ All notable changes to this project will be documented in this file.
 
 **Warning:** Features marked as *experimental* may change or be removed in a future release without notice. Use with caution.
 
-## [Unreleased]
+<!-- ## [Unreleased] -->
+
+## [2.0.0-beta.2]
 
 ### Added
 
@@ -16,6 +18,7 @@ All notable changes to this project will be documented in this file.
 
 * The HTTP server now requests that publication resources are not cached by browsers.
   * Caching poses a security risk for protected publications.
+
 
 ## [2.0.0-beta.1]
 
@@ -81,4 +84,5 @@ server.loadCustomResource(assets.open("scripts/highlight.js"), "highlight.js", I
 [2.0.0-alpha.1]: https://github.com/readium/r2-streamer-kotlin/compare/1.1.5...2.0.0-alpha.1
 [2.0.0-alpha.2]: https://github.com/readium/r2-streamer-kotlin/compare/2.0.0-alpha.1...2.0.0-alpha.2
 [2.0.0-beta.1]: https://github.com/readium/r2-streamer-kotlin/compare/2.0.0-alpha.2...2.0.0-beta.1
+[2.0.0-beta.2]: https://github.com/readium/r2-streamer-kotlin/compare/2.0.0-beta.1...2.0.0-beta.2